python-app/requirements.txt
aiohttp
asyncio
bs4
Flask
flask-cors
google-genai
pymongo
requests
selenium
setuptools
transformers
urllib3
dotenv
tldextract